增加浮点数验证和单边范围验证

+++ src/org/xresloader/core/data/vfy/DataVerifyResult.java
@@ -2,5 +2,5 @@ package org.xresloader.core.data.vfy;
 
 public class DataVerifyResult {
     public boolean success = false;
-    public long value = 0;
+    public double value = 0;
 }
